Fix: Allow fallthrough comment in case block (fixes #9559)
The rule `no-fallthrough` now gets the `SwitchCase` that falls through passed into the function where the comment is found and checked. In that function, if the falllthrough case has a block, it looks for a matching comment inside the block.

Also added are corresponding tests and an example in the documentation for the switch case block.